+++ ConsoleBalatro/Engine/Events/Args/EngineHandPlayDoneArgs.cs
@@ -12,8 +12,8 @@ namespace ConsoleBalatro.Engine.Events.Args
         public PlayedHandType HandTypeThatWasPlayed { get; set; }
         public List<Card> CardsInPlayedHand { get; set; }
         public List<Card> CardsHeldInHand { get; set; }
-        public int CurrentTotalChips { get; set; }
-        public int RequiredChipsForBlind { get; set; }
+        public long CurrentTotalChips { get; set; }
+        public long RequiredChipsForBlind { get; set; }
         public bool PreventGameOverAndWinBlind { get; set; } = false;
 
     }
